Add custom launchd log - mac unified logging sucks
While trying to help people for whom the launchd job failed, we found it
very difficult to find the entries in the new(ish) MacOS unified logging
thing, so update the launchd job to write output to a temporary log file.

The file is intentionally saved in /tmp and overwritten each time the
launchd job runs to avoid having to deal with permissions or rotation.

File name matches the plist name, e.g. homebrew.mxcl.aws-rotate-iam-keys.log