Allow homebrew install to skip awscli, python etc.
Installing via homebrew on MacOS automatically upgrades awscli to the
latest stable version, because the homebrew formula depends on awscli,
and homebrew thinks it's wise to automatically upgrade dependencies.

This might seem sensible in theory, but in practice can cause chaos
because a simple install of aws-rotate-iam-keys can cause both python
and readline to be upgraded/replaced (because awscli depends on python,
and python depends on readline), which breaks anything dependent on the
previously installed versions (e.g. any python packages installed in
develop mode, or any homebrew installed binaries dependent on readline,
e.g. gawk, and other things dependent on those, e.g. bash completion).

Arguably, these are upstream issues, or just symptomatic of how homebrew
is designed to work, and we shouldn't worry about them here; but this is
not the first time a reinstall of aws-rotate-iam-keys has temporarily
crippled my machine, and it would be nice to stop that happening again.

This commit allows that, in the least intrusive way I could find, by
adding an option to `brew install aws-rotate-iam-keys --without-awscli`.