Check profiles have the same key before rotating
We need some way to verify multiple profile arguments refer to the same
IAM user, otherwise it's possible to accidentally break the credentials
configuration by creating a new key for one IAM user and writing that to
a profile that's currently using the access keys for another IAM user.

There might be occasions where this is the intended behaviour, but it
seems that's very likely to be the edge rather than the common case, so
just check the access key ids for all profiles match before continuing.

Note: in future there could be a -f/--force option to allow overriding