Allow installation of unstable, forked, versions
Add head and devel specs to the homebrew formula to allow installation
from the master and develop branches respectively of whichever repo is
the source of the homebrew tap. This is fugly, but it seems a necessary
evil to allow us to maintain a usable fork including new features not
yet merged upstream, without breaking upstream compatibility.

Yeah, I know, it's a horrible hack, I hate it too, but it does the job.

P.S. Yes, this will prevent the formula being accepted to Homebrew core,
but that's not going to happen anytime soon, so not worth worrying about