Consistently use for; do rather than for \n do
Doesn't really matter what the rule is, as long as it's applied
consistently, and previously we had a mix of both. The choice of
for; do over for \n do is based purely on the current consistent
application of if; then and while; do rather than if \n then etc.