Fix homebrew devel version (don't set to git ref)
The previous code attempted to show the version of aws-rotate-iam-keys
that would be installed as the git ref (typically a shortened sha ref).

This code tried to `git describe` the develop branch on the repo, but a
develop branch doesn't exist on the main rhyeal repo (only on my fork),
which caused a warning "fatal: Not a valid object name develop" whenever
the formula was loaded (e.g. during installation). Although this didn't
prevent installation, it was quite alarming, and needed to be fixed.

The version was explicitly set because it could not be inferred from the
git url, with the aim of making it clear which version `--devel` would
install. However, it turns out this was not accurate, because Homebrew
would automatically pull changes on install, thereby installing the
latest develop (potentially ahead the local develop added by the tap).

To both fix the warning, and work around the inaccurate version info,
I've updated the version to the string "head", which should be clear
enough (to git users), and does not require the develop branch to exist.

Note that "HEAD" is a magic word/version in the land of Homebrew, hence
the lower case "head". Using "HEAD" causes strange things to happen.