Fix path to aws-rotate-iam-keys in debian crontab
The debian package post install script updates the user's crontab to add
a job to run aws-rotate-iam-keys and rotate keys for the default profile.

The path to the program after installation seems dependent on the distro
though, /bin on debian itself, but /usr/bin on ubuntu. This causes the
crontab entry added on debian to fail, because the path to the program
is hard-coded as /usr/bin/aws-rotate-iam-keys. Fix by updating the post
install and pre remove scripts to try detecting the path using the which
command, and falling back to /usr/bin/aws-rotate-iam-keys if that fails.

Although I haven't fully tested this, as doing so would require building
a test debian package and installing it, I'm pretty sure it'll work ok,
on the assumption that $PATH at install time includes /bin and /usr/bin.