Disable AWS CLI pager to avoid user interaction
This shouldn't happen when the script is run from cron or homebrew
services, because a pager should only be used when stdout is a tty.

However, issue #50 suggests that there might be an AWS CLI bug that
results in the pager being used even when stdout is not a tty, and
even if that's not the case, we don't want the pager running anyway.