Remove Homebrew formula template, update manually
Copy recent changes to Homebrew formula template by @Volatus to actual
formula file and delete the template. CircleCI builds are broken, and
using a template for this was probably a bad idea anyway.

This does not update the latest stable release, it remains at 0.9.8.4

To install the latest version, use `brew install --HEAD`